% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entcn.R
\name{translate}
\alias{translate}
\title{Translate English words into Chinese words}
\usage{
translate(x)
}
\arguments{
\item{x}{A English words or a vector made up of English words}
}
\value{
The Chinese words equivalent of English words
}
\description{
When you pass in an English words or a vector consisting of an English words,
this function will calls the Youdao translation open API for R to return the
corresponding type of Chinese representation.
}
\examples{
# Example-1
translate("English")

# Example-2
x <- c("father", "mother", "son")
translate(x)

}
